Script to output team utilization and overlap of teams between players.
This script analyzes the team data that players in the competition have chosen; two files are output.  One is a CSV file with one row per possible pairings of entries and contains the teams that both teams have, along with the size of that set.  The other contains a list of teams with the number and percentage of entries that contain that team.